refactor(attempt-review): Merge page content into review page
